+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 1.1 (Unreleased)
 * Connected Recognition Awards are handled in a special way.
 * No longer verify SSL when checking Github for updated version.
+* Script asks for any overtime/on-call pay.
 
 ### 1.0 (May 3rd, 2013)